MCP Client: Move token exchange customizer to a sample
Per the discussion in #88, hold off on integrating token exchange into
mcp-client-security, and ship it as a sample instead. The customizer
moves to a new samples/sample-mcp-client-token-exchange module, with
the Keycloak-compatible subject token resolver (sending the subject
token as an access token rather than a JWT) bundled via a factory
method, and a README covering the resource-server topology, Keycloak
notes, and initialization guidance.

See #88
