+++ pom.xml
@@ -29,6 +29,7 @@
         <module>samples/sample-mcp-server-api-key</module>
         <module>samples/sample-mcp-client</module>
         <module>samples/sample-mcp-client-webclient</module>
+        <module>samples/sample-mcp-client-token-exchange</module>
         <module>samples/integration-tests</module>
     </modules>
